OvmfPkg: Update IoMmuDxe to support both TDX and SEV
